+++ joe/utils.h
@@ -43,7 +43,7 @@ extern void *ITEM; /* Temporary global variable (from queue.c) */
 
 /* Allocate item from free-list.  If free-list empty, replenish it. */
 
-void *replenish PARAMS((void **list,int size));
+void *replenish PARAMS((void **list,size_t size));
 
 #define al_single(list,type) ( \
 	(ITEM = *(void **)(list)) ? \
@@ -78,7 +78,7 @@ struct zs {
 	gc_add(&gc, &(var), rm_zs); \
 } while(0)
 
-ZS raw_mk_zs PARAMS((GC **gc,unsigned char *s,int len));
+ZS raw_mk_zs PARAMS((GC **gc,unsigned char *s,size_t len));
 
 /* Destructor for zs */
 
@@ -139,7 +139,7 @@ int joe_set_signal PARAMS((int signum, sighandler_t handler));
 
 /* Simple parsers */
 int parse_ws PARAMS((unsigned char **p,int cmt));
-int parse_ident PARAMS((unsigned char **p,unsigned char *buf,int len));
+int parse_ident PARAMS((unsigned char **p,unsigned char *buf,size_t len));
 int parse_kw PARAMS((unsigned char **p,unsigned char *kw));
 long parse_num PARAMS((unsigned char **p));
 int parse_tows PARAMS((unsigned char **p,unsigned char *buf));
@@ -147,8 +147,8 @@ int parse_field PARAMS((unsigned char **p,unsigned char *field));
 int parse_char PARAMS((unsigned char  **p,unsigned char c));
 int parse_int PARAMS((unsigned char **p,int *buf));
 int parse_long PARAMS((unsigned char **p,long  *buf));
-int parse_string PARAMS((unsigned char **p,unsigned char *buf,int len));
+ssize_t parse_string PARAMS((unsigned char **p,unsigned char *buf,size_t len));
 int parse_range PARAMS((unsigned char **p,int *first,int *second));
-void emit_string PARAMS((FILE *f,unsigned char *s,int len));
+void emit_string PARAMS((FILE *f,unsigned char *s,size_t len));
 
 #endif
