+++ joe/va.c
@@ -7,9 +7,9 @@
  */
 #include "types.h"
 
-aELEMENT *vamk(int len)
+aELEMENT *vamk(size_t len)
 {
-	int *new = (int *) joe_malloc((1 + len) * sizeof(aELEMENT) + 2 * sizeof(int));
+	size_t *new = (size_t *) joe_malloc((1 + len) * sizeof(aELEMENT) + 2 * sizeof(size_t));
 
 	new[0] = len;
 	new[1] = 0;
@@ -21,38 +21,38 @@ void varm(aELEMENT *vary)
 {
 	if (vary) {
 		vazap(vary, 0, aLen(vary));
-		joe_free((int *) vary - 2);
+		joe_free((size_t *) vary - 2);
 	}
 }
 
-int alen(aELEMENT *ary)
+size_t alen(aELEMENT *ary)
 {
 	if (ary) {
 		aELEMENT *beg = ary;
 		while (acmp(*ary, aterm))
 			++ary;
-		return ary - beg;
+		return (size_t)(ary - beg);
 	} else
 		return 0;
 }
 
-aELEMENT *vaensure(aELEMENT *vary, int len)
+aELEMENT *vaensure(aELEMENT *vary, size_t len)
 {
 	if (!vary)
 		vary = vamk(len);
 	else if (len > aSiz(vary)) {
 		len += (len >> 2);
-		vary = (aELEMENT *)(2 + (int *) joe_realloc((int *) vary - 2, (len + 1) * sizeof(aELEMENT) + 2 * sizeof(int)));
+		vary = (aELEMENT *)(2 + (size_t *) joe_realloc((size_t *) vary - 2, (len + 1) * sizeof(aELEMENT) + 2 * sizeof(size_t)));
 
 		aSiz(vary) = len;
 	}
 	return vary;
 }
 
-aELEMENT *vazap(aELEMENT *vary, int pos, int n)
+aELEMENT *vazap(aELEMENT *vary, size_t pos, size_t n)
 {
 	if (vary) {
-		int x;
+		size_t x;
 
 		if (pos < aLen(vary)) {
 			if (pos + n <= aLen(vary)) {
@@ -67,7 +67,7 @@ aELEMENT *vazap(aELEMENT *vary, int pos, int n)
 	return vary;
 }
 
-aELEMENT *vatrunc(aELEMENT *vary, int len)
+aELEMENT *vatrunc(aELEMENT *vary, size_t len)
 {
 	if (!vary || len > aLEN(vary))
 		vary = vaensure(vary, len);
@@ -81,9 +81,9 @@ aELEMENT *vatrunc(aELEMENT *vary, int len)
 	return vary;
 }
 
-aELEMENT *vafill(aELEMENT *vary, int pos, aELEMENT el, int len)
+aELEMENT *vafill(aELEMENT *vary, size_t pos, aELEMENT el, size_t len)
 {
-	int olen = aLEN(vary), x;
+	size_t olen = aLEN(vary), x;
 
 	if (!vary || pos + len > aSIZ(vary))
 		vary = vaensure(vary, pos + len);
@@ -99,9 +99,9 @@ aELEMENT *vafill(aELEMENT *vary, int pos, aELEMENT el, int len)
 }
 
 #ifdef junk
-aELEMENT *vancpy(aELEMENT *vary, int pos, aELEMENT *array, int len)
+aELEMENT *vancpy(aELEMENT *vary, size_t pos, aELEMENT *array, size_t len)
 {
-	int olen = aLEN(vary);
+	size_t olen = aLEN(vary);
 
 	if (!vary || pos + len > aSIZ(vary))
 		vary = vaensure(vary, pos + len);
@@ -116,9 +116,9 @@ aELEMENT *vancpy(aELEMENT *vary, int pos, aELEMENT *array, int len)
 }
 #endif
 
-aELEMENT *vandup(aELEMENT *vary, int pos, aELEMENT *array, int len)
+aELEMENT *vandup(aELEMENT *vary, size_t pos, aELEMENT *array, size_t len)
 {
-	int olen = aLEN(vary), x;
+	size_t olen = aLEN(vary), x;
 
 	if (!vary || pos + len > aSIZ(vary))
 		vary = vaensure(vary, pos + len);
@@ -138,7 +138,7 @@ aELEMENT *vadup(aELEMENT *vary)
 	return vandup(NULL, 0, vary, aLEN(vary));
 }
 
-aELEMENT *_vaset(aELEMENT *vary, int pos, aELEMENT el)
+aELEMENT *_vaset(aELEMENT *vary, size_t pos, aELEMENT el)
 {
 	if (!vary || pos + 1 > aSIZ(vary))
 		vary = vaensure(vary, pos + 1);
@@ -163,7 +163,7 @@ static int _acmp(aELEMENT *a, aELEMENT *b)
 	return acmp(*a, *b);
 }
 
-aELEMENT *vasort(aELEMENT *ary, int len)
+aELEMENT *vasort(aELEMENT *ary, size_t len)
 {
 	if (!ary || !len)
 		return ary;
@@ -171,9 +171,9 @@ aELEMENT *vasort(aELEMENT *ary, int len)
 	return ary;
 }
 
-aELEMENT *vawords(aELEMENT *a, unsigned char *s, int len, unsigned char *sep, int seplen)
+aELEMENT *vawords(aELEMENT *a, unsigned char *s, size_t len, unsigned char *sep, size_t seplen)
 {
-	int x;
+	size_t x;
 
 	if (!a)
 		a = vamk(10);
@@ -185,7 +185,7 @@ aELEMENT *vawords(aELEMENT *a, unsigned char *s, int len, unsigned char *sep, in
 	len -= x;
 	if (len) {
 		x = vsscan(s, len, sep, seplen);
-		if (x != ~0) {
+		if (x != ~(size_t)0) {
 			a = vaadd(a, vsncpy(vsmk(x), 0, s, x));
 			s += x;
 			len -= x;
